The translator also copies the '.row' file if available.
Added back support for v4 CFG files. Simulator sets
mem_buses = # processors and 1 half duplex intra-node link
when using v4 configuration files.

When copying the output PCF, all information befor the event
types are embedded, using the Dimemas default states

+++ Simulator/model/communic.h
@@ -77,7 +77,7 @@ void COMMUNIC_wait(struct t_thread *thread);
 void COMMUNIC_block_after_busy_wait(struct t_thread *thread);
 
 /* It is used also in 'memory' and 'ports' */
-void transferencia(int size,
+void transferencia(long long int size,
                    t_boolean remote,
                    struct t_thread *thread,
                    struct t_dedicated_connection *connection,
